feat: add advanced OAuth2 parameters management and grant types handling
