+++ packages/hoppscotch-common/src/components/http/authorization/OAuth2.vue
@@ -541,16 +541,11 @@
 <script setup lang="ts">
 import { HoppGQLAuthOAuth2, HoppRESTAuthOAuth2 } from "@hoppscotch/data"
 import { useService } from "dioc/vue"
-import * as E from "fp-ts/Either"
-import { Ref, computed, ref, watch, onMounted } from "vue"
-import { z } from "zod"
+import { computed, onMounted, ref } from "vue"
 import { useI18n } from "~/composables/i18n"
-import { refWithCallbackOnChange } from "~/composables/ref"
+import { useOAuth2AdvancedParams } from "~/composables/oauth2/useOAuth2AdvancedParams"
+import { useOAuth2GrantTypes } from "~/composables/oauth2/useOAuth2GrantTypes"
 import { useToast } from "~/composables/toast"
-import {
-  replaceTemplateString,
-  replaceTemplateStringsInObjectValues,
-} from "~/helpers/auth"
 import {
   commonOAuth2AuthParams,
   commonOAuth2RefreshParams,
@@ -559,51 +554,22 @@ import {
   sendInOptionsLabels,
 } from "~/helpers/oauth2Params"
 import { AggregateEnvironment } from "~/newstore/environments"
-import authCode, {
-  AuthCodeOauthFlowParams,
-  AuthCodeOauthRefreshParams,
-  getDefaultAuthCodeOauthFlowParams,
-} from "~/services/oauth/flows/authCode"
-import clientCredentials, {
-  ClientCredentialsFlowParams,
-  getDefaultClientCredentialsFlowParams,
-} from "~/services/oauth/flows/clientCredentials"
-import implicit, {
-  ImplicitOauthFlowParams,
-  getDefaultImplicitOauthFlowParams,
-} from "~/services/oauth/flows/implicit"
-import passwordFlow, {
-  PasswordFlowParams,
-  getDefaultPasswordFlowParams,
-} from "~/services/oauth/flows/password"
 import {
-  PersistedOAuthConfig,
   grantTypesInvolvingRedirect,
+  PersistedOAuthConfig,
 } from "~/services/oauth/oauth.service"
+import * as E from "fp-ts/Either"
 import { PersistenceService } from "~/services/persistence"
 import { GQLTabService } from "~/services/tab/graphql"
 import { RESTTabService } from "~/services/tab/rest"
-import IconCircle from "~icons/lucide/circle"
-import IconCircleDot from "~icons/lucide/circle-dot"
 import IconChevronDown from "~icons/lucide/chevron-down"
 import IconChevronUp from "~icons/lucide/chevron-up"
+import IconCircle from "~icons/lucide/circle"
+import IconCircleDot from "~icons/lucide/circle-dot"
 import IconPlus from "~icons/lucide/plus"
 
 const t = useI18n()
 const toast = useToast()
-
-const gqlTabsService = useService(GQLTabService)
-const persistenceService = useService(PersistenceService)
-const restTabsService = useService(RESTTabService)
-
-const props = defineProps<{
-  modelValue: HoppRESTAuthOAuth2 | HoppGQLAuthOAuth2
-  isCollectionProperty?: boolean
-  envs?: AggregateEnvironment[]
-  source: "REST" | "GraphQL"
-}>()
-
-const auth = ref(props.modelValue)
 
 const addToTargets = [
   {
@@ -616,6 +582,80 @@ const addToTargets = [
   },
 ]
 
+const getPlaceholderForField = (fieldId: string): string => {
+  const placeholders: Record<string, string> = {
+    authEndpoint: "https://example.com/oauth2/authorize",
+    tokenEndpoint: "https://example.com/oauth2/token",
+    clientId: "your_client_id_here",
+    clientSecret: "your_client_secret_here",
+    scopes: "read write",
+    username: "your_username",
+    password: "your_password",
+  }
+  return placeholders[fieldId] || t("authorization.oauth.enter_value")
+}
+
+const props = defineProps<{
+  modelValue: HoppRESTAuthOAuth2 | HoppGQLAuthOAuth2
+  isCollectionProperty?: boolean
+  envs?: AggregateEnvironment[]
+  source: "REST" | "GraphQL"
+}>()
+
+const auth = ref(props.modelValue)
+
+// Advanced Configuration state
+const isAdvancedConfigExpanded = ref(false)
+
+const toggleAdvancedConfig = () => {
+  isAdvancedConfigExpanded.value = !isAdvancedConfigExpanded.value
+}
+
+// Get advanced params management
+const {
+  workingAuthRequestParams,
+  workingTokenRequestParams,
+  workingRefreshRequestParams,
+  addAuthRequestParam,
+  updateAuthRequestParam,
+  deleteAuthRequestParam,
+  addTokenRequestParam,
+  updateTokenRequestParam,
+  deleteTokenRequestParam,
+  addRefreshRequestParam,
+  updateRefreshRequestParam,
+  deleteRefreshRequestParam,
+  initializeParams,
+} = useOAuth2AdvancedParams(auth)
+
+// Initialize advanced params on mount. e.g. authRequestParams, tokenRequestParams, refreshRequestParams
+onMounted(initializeParams)
+
+// Template refs for tippy actions
+const grantTypeTippyActions = ref<HTMLElement | null>(null)
+const pkceTippyActions = ref<HTMLElement | null>(null)
+const authTippyActions = ref<HTMLElement | null>(null)
+const clientAuthenticationTippyActions = ref<HTMLElement | null>(null)
+
+// Get grant types management
+const {
+  supportedGrantTypes,
+  selectedGrantTypeID,
+  selectedGrantType,
+  changeSelectedGrantType,
+  runAction,
+  runTokenRefresh,
+  currentOAuthGrantTypeFormElements,
+} = useOAuth2GrantTypes(
+  auth,
+  (...args) => setAccessTokenInActiveContext(...args),
+  workingAuthRequestParams,
+  workingTokenRequestParams,
+  workingRefreshRequestParams,
+  pkceTippyActions,
+  clientAuthenticationTippyActions
+)
+
 const passBy = computed(() => {
   return (
     addToTargets.find((target) => target.id === auth.value.addTo)?.label ||
@@ -623,731 +663,9 @@ const passBy = computed(() => {
   )
 })
 
-const supportedGrantTypes = [
-  {
-    // used for both authCode and authCodePKCE
-    id: "authCode" as const,
-    label: t("authorization.oauth.label_auth_code"),
-    formElements: computed(() => {
-      if (!(auth.value.grantTypeInfo.grantType === "AUTHORIZATION_CODE")) {
-        return
-      }
-
-      const grantType = auth.value.grantTypeInfo
-
-      const authEndpoint = refWithCallbackOnChange(
-        grantType?.authEndpoint,
-        (value) => {
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            authEndpoint: value,
-          }
-        }
-      )
-
-      const tokenEndpoint = refWithCallbackOnChange(
-        grantType?.tokenEndpoint,
-        (value) => {
-          if (!("tokenEndpoint" in auth.value.grantTypeInfo)) {
-            return
-          }
-
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            tokenEndpoint: value,
-          }
-        }
-      )
-
-      const clientID = refWithCallbackOnChange(grantType?.clientID, (value) => {
-        auth.value.grantTypeInfo = {
-          ...auth.value.grantTypeInfo,
-          clientID: value,
-        }
-      })
-
-      const clientSecret = refWithCallbackOnChange(
-        grantType?.clientSecret,
-        (value) => {
-          if (!("clientSecret" in auth.value.grantTypeInfo)) {
-            return
-          }
-
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            clientSecret: value ?? "",
-          }
-        }
-      )
-
-      const scopes = refWithCallbackOnChange(
-        grantType?.scopes ? grantType.scopes : undefined,
-        (value) => {
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            scopes: value,
-          }
-        }
-      )
-
-      const isPKCE = refWithCallbackOnChange(
-        auth.value.grantTypeInfo.isPKCE,
-        (value) => {
-          if (!("isPKCE" in auth.value.grantTypeInfo)) {
-            return
-          }
-
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            isPKCE: value,
-          }
-        }
-      )
-
-      const codeChallenge: Ref<{
-        id: "plain" | "S256"
-        label: string
-      } | null> = refWithCallbackOnChange(
-        auth.value.grantTypeInfo.codeVerifierMethod
-          ? {
-              id: auth.value.grantTypeInfo.codeVerifierMethod,
-              label:
-                auth.value.grantTypeInfo.codeVerifierMethod === "plain"
-                  ? "Plain"
-                  : "SHA-256",
-            }
-          : null,
-        (value) => {
-          if (!("codeVerifierMethod" in auth.value.grantTypeInfo) || !value) {
-            return
-          }
-
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            codeVerifierMethod: value.id,
-          }
-        }
-      )
-
-      const refreshToken = async () => {
-        const grantTypeInfo = auth.value.grantTypeInfo
-
-        console.log("grantTypeInfo", grantTypeInfo)
-
-        if (!("refreshToken" in grantTypeInfo)) {
-          return E.left("NO_REFRESH_TOKEN_PRESENT" as const)
-        }
-
-        const refreshToken = grantTypeInfo.refreshToken
-
-        if (!refreshToken) {
-          return E.left("NO_REFRESH_TOKEN_PRESENT" as const)
-        }
-
-        const params: AuthCodeOauthRefreshParams = {
-          clientID: clientID.value,
-          clientSecret: clientSecret.value,
-          tokenEndpoint: tokenEndpoint.value,
-          refreshToken,
-        }
-
-        const unwrappedParams = replaceTemplateStringsInObjectValues(params)
-
-        const refreshTokenFunc = authCode.refreshToken
-
-        if (!refreshTokenFunc) {
-          return E.left("REFRESH_TOKEN_FUNCTION_NOT_DEFINED" as const)
-        }
-
-        const res = await refreshTokenFunc(unwrappedParams)
-
-        if (E.isLeft(res)) {
-          return E.left("OAUTH_REFRESH_TOKEN_FAILED" as const)
-        }
-
-        setAccessTokenInActiveContext(
-          res.right.access_token,
-          res.right.refresh_token
-        )
-
-        return E.right(undefined)
-      }
-
-      const runAction = async () => {
-        const params: AuthCodeOauthFlowParams = {
-          authEndpoint: authEndpoint.value,
-          tokenEndpoint: tokenEndpoint.value,
-          clientID: clientID.value,
-          clientSecret: clientSecret.value,
-          scopes: scopes.value,
-          isPKCE: isPKCE.value,
-          codeVerifierMethod: codeChallenge.value?.id,
-          authRequestParams: workingAuthRequestParams.value
-            .filter((p) => p.active && p.key && p.value)
-            .map((p) => ({
-              id: p.id,
-              key: replaceTemplateString(p.key),
-              value: replaceTemplateString(p.value),
-              active: p.active,
-              sendIn: p.sendIn,
-            })),
-          tokenRequestParams: workingTokenRequestParams.value
-            .filter((p) => p.active && p.key && p.value)
-            .map((p) => ({
-              id: p.id,
-              key: replaceTemplateString(p.key),
-              value: replaceTemplateString(p.value),
-              active: p.active,
-              sendIn: p.sendIn,
-            })),
-          refreshRequestParams: workingRefreshRequestParams.value
-            .filter((p) => p.active && p.key && p.value)
-            .map((p) => ({
-              id: p.id,
-              key: replaceTemplateString(p.key),
-              value: replaceTemplateString(p.value),
-              active: p.active,
-              sendIn: p.sendIn,
-            })),
-        }
-
-        const unwrappedParams = replaceTemplateStringsInObjectValues(params)
-
-        const parsedArgs = authCode.params.safeParse(unwrappedParams)
-
-        if (!parsedArgs.success) {
-          return E.left("VALIDATION_FAILED" as const)
-        }
-
-        const res = await authCode.init(parsedArgs.data)
-
-        if (E.isLeft(res)) {
-          return res
-        }
-
-        return E.right(undefined)
-      }
-
-      const pkceElements = computed(() => {
-        const checkbox = {
-          id: "isPKCE",
-          label: t("authorization.oauth.label_use_pkce"),
-          type: "checkbox" as const,
-          ref: isPKCE,
-          onChange: (e: Event) => {
-            const target = e.target as HTMLInputElement
-
-            isPKCE.value = target.checked
-          },
-        }
-
-        return isPKCE.value
-          ? [
-              checkbox,
-              {
-                id: "codeChallenge",
-                label: t("authorization.oauth.label_code_challenge"),
-                type: "dropdown" as const,
-                ref: codeChallenge,
-                tippyRefName: "pkceTippyActions",
-                tippyRef: pkceTippyActions,
-                options: [
-                  {
-                    id: "plain" as const,
-                    label: "Plain",
-                  },
-                  {
-                    id: "S256" as const,
-                    label: "SHA-256",
-                  },
-                ],
-              },
-            ]
-          : [checkbox]
-      })
-
-      const elements = computed(() => {
-        return [
-          ...pkceElements.value,
-          {
-            id: "authEndpoint",
-            label: t("authorization.oauth.label_authorization_endpoint"),
-            type: "text" as const,
-            ref: authEndpoint,
-          },
-          {
-            id: "tokenEndpoint",
-            label: t("authorization.oauth.label_token_endpoint"),
-            type: "text" as const,
-            ref: tokenEndpoint,
-          },
-          {
-            id: "clientId",
-            label: t("authorization.oauth.label_client_id"),
-            type: "text" as const,
-            ref: clientID,
-          },
-          {
-            id: "clientSecret",
-            label: t("authorization.oauth.label_client_secret"),
-            type: "text" as const,
-            ref: clientSecret,
-          },
-          {
-            id: "scopes",
-            label: t("authorization.oauth.label_scopes"),
-            type: "text" as const,
-            ref: scopes,
-          },
-        ]
-      })
-
-      return {
-        runAction,
-        refreshToken,
-        elements,
-      }
-    }),
-  },
-  {
-    id: "clientCredentials" as const,
-    label: t("authorization.oauth.label_client_credentials"),
-    formElements: computed(() => {
-      if (!(auth.value.grantTypeInfo.grantType === "CLIENT_CREDENTIALS")) {
-        return
-      }
-
-      const grantTypeInfo = auth.value.grantTypeInfo
-
-      const authEndpoint = refWithCallbackOnChange(
-        grantTypeInfo?.authEndpoint,
-        (value) => {
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            authEndpoint: value,
-          }
-        }
-      )
-
-      const clientID = refWithCallbackOnChange(
-        grantTypeInfo?.clientID,
-        (value) => {
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            clientID: value,
-          }
-        }
-      )
-
-      const clientSecret = refWithCallbackOnChange(
-        grantTypeInfo?.clientSecret,
-        (value) => {
-          if (!("clientSecret" in auth.value.grantTypeInfo)) {
-            return
-          }
-
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            clientSecret: value,
-          }
-        }
-      )
-
-      const scopes = refWithCallbackOnChange(
-        grantTypeInfo?.scopes ? grantTypeInfo.scopes : undefined,
-        (value) => {
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            scopes: value,
-          }
-        }
-      )
-
-      const clientAuthentication = refWithCallbackOnChange(
-        grantTypeInfo.clientAuthentication
-          ? grantTypeInfo.clientAuthentication === "AS_BASIC_AUTH_HEADERS"
-            ? {
-                id: "AS_BASIC_AUTH_HEADERS" as const,
-                label: t("authorization.oauth.label_send_as_basic_auth"),
-              }
-            : {
-                id: "IN_BODY" as const,
-                label: t("authorization.oauth.label_send_in_body"),
-              }
-          : {
-              id: "IN_BODY" as const,
-              label: t("authorization.oauth.label_send_in_body"),
-            },
-        (value) => {
-          if (!("clientAuthentication" in auth.value.grantTypeInfo)) {
-            return
-          }
-
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            clientAuthentication: value.id,
-          }
-        }
-      )
-
-      const runAction = async () => {
-        const values: ClientCredentialsFlowParams =
-          replaceTemplateStringsInObjectValues({
-            authEndpoint: authEndpoint.value,
-            clientID: clientID.value,
-            clientSecret: clientSecret.value,
-            scopes: scopes.value,
-            clientAuthentication: clientAuthentication.value.id,
-          })
-
-        const parsedArgs = clientCredentials.params.safeParse(values)
-
-        if (!parsedArgs.success) {
-          return E.left("VALIDATION_FAILED" as const)
-        }
-
-        const res = await clientCredentials.init(parsedArgs.data)
-
-        if (E.isLeft(res)) {
-          return E.left("OAUTH_TOKEN_FETCH_FAILED" as const)
-        }
-
-        setAccessTokenInActiveContext(res.right?.access_token)
-
-        toast.success(t("authorization.oauth.token_fetched_successfully"))
-
-        return E.right(undefined)
-      }
-
-      const elements = computed(() => {
-        return [
-          {
-            id: "authEndpoint",
-            label: t("authorization.oauth.label_authorization_endpoint"),
-            type: "text" as const,
-            ref: authEndpoint,
-          },
-          {
-            id: "clientId",
-            label: t("authorization.oauth.label_client_id"),
-            type: "text" as const,
-            ref: clientID,
-          },
-          {
-            id: "clientSecret",
-            label: t("authorization.oauth.label_client_secret"),
-            type: "text" as const,
-            ref: clientSecret,
-          },
-          {
-            id: "scopes",
-            label: t("authorization.oauth.label_scopes"),
-            type: "text" as const,
-            ref: scopes,
-          },
-          {
-            id: "clientAuthentication",
-            label: t("authorization.oauth.label_send_as"),
-            type: "dropdown" as const,
-            ref: clientAuthentication,
-            tippyRefName: "clientAuthenticationTippyActions",
-            tippyRef: clientAuthenticationTippyActions,
-            options: [
-              {
-                id: "IN_BODY" as const,
-                label: t("authorization.oauth.label_send_in_body"),
-              },
-              {
-                id: "AS_BASIC_AUTH_HEADERS" as const,
-                label: t("authorization.oauth.label_send_as_basic_auth"),
-              },
-            ],
-          },
-        ]
-      })
-
-      return {
-        runAction,
-        elements,
-      }
-    }),
-  },
-  {
-    id: "password" as const,
-    label: "Password",
-    formElements: computed(() => {
-      if (!(auth.value.grantTypeInfo.grantType === "PASSWORD")) {
-        return
-      }
-
-      const grantTypeInfo = auth.value.grantTypeInfo
-
-      const authEndpoint = refWithCallbackOnChange(
-        grantTypeInfo?.authEndpoint,
-        (value) => {
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            authEndpoint: value,
-          }
-        }
-      )
-
-      const clientID = refWithCallbackOnChange(
-        grantTypeInfo?.clientID,
-        (value) => {
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            clientID: value,
-          }
-        }
-      )
-
-      const clientSecret = refWithCallbackOnChange(
-        grantTypeInfo?.clientSecret,
-        (value) => {
-          if (!("clientSecret" in auth.value.grantTypeInfo)) {
-            return
-          }
-
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            clientSecret: value,
-          }
-        }
-      )
-
-      const scopes = refWithCallbackOnChange(
-        grantTypeInfo?.scopes ? grantTypeInfo.scopes : undefined,
-        (value) => {
-          auth.value.grantTypeInfo.scopes = value
-        }
-      )
-
-      const username = refWithCallbackOnChange(
-        grantTypeInfo?.username,
-        (value) => {
-          if (!("username" in auth.value.grantTypeInfo)) {
-            return
-          }
-
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            username: value,
-          }
-        }
-      )
-
-      const password = refWithCallbackOnChange(
-        grantTypeInfo?.password,
-        (value) => {
-          if (!("password" in auth.value.grantTypeInfo)) {
-            return
-          }
-
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            password: value,
-          }
-        }
-      )
-
-      const runAction = async () => {
-        const values: PasswordFlowParams = replaceTemplateStringsInObjectValues(
-          {
-            authEndpoint: authEndpoint.value,
-            clientID: clientID.value,
-            clientSecret: clientSecret.value,
-            scopes: scopes.value,
-            username: username.value,
-            password: password.value,
-          }
-        )
-
-        const parsedArgs = passwordFlow.params.safeParse(values)
-
-        if (!parsedArgs.success) {
-          return E.left("VALIDATION_FAILED" as const)
-        }
-
-        const res = await passwordFlow.init(parsedArgs.data)
-
-        if (E.isLeft(res)) {
-          return E.left("OAUTH_TOKEN_FETCH_FAILED" as const)
-        }
-
-        setAccessTokenInActiveContext(res.right?.access_token)
-
-        toast.success(t("authorization.oauth.token_fetched_successfully"))
-
-        return E.right(undefined)
-      }
-
-      const elements = computed(() => {
-        return [
-          {
-            id: "authEndpoint",
-            label: t("authorization.oauth.label_authorization_endpoint"),
-            type: "text" as const,
-            ref: authEndpoint,
-          },
-          {
-            id: "clientId",
-            label: t("authorization.oauth.label_client_id"),
-            type: "text" as const,
-            ref: clientID,
-          },
-          {
-            id: "clientSecret",
-            label: t("authorization.oauth.label_client_secret"),
-            type: "text" as const,
-            ref: clientSecret,
-          },
-          {
-            id: "username",
-            label: t("authorization.oauth.label_username"),
-            type: "text" as const,
-            ref: username,
-          },
-          {
-            id: "password",
-            label: t("authorization.oauth.label_password"),
-            type: "text" as const,
-            ref: password,
-          },
-          {
-            id: "scopes",
-            label: t("authorization.oauth.label_scopes"),
-            type: "text" as const,
-            ref: scopes,
-          },
-        ]
-      })
-
-      return {
-        runAction,
-        elements,
-      }
-    }),
-  },
-  {
-    id: "implicit" as const,
-    label: t("authorization.oauth.label_implicit"),
-    formElements: computed(() => {
-      const grantTypeInfo = auth.value.grantTypeInfo
-
-      const authEndpoint = refWithCallbackOnChange(
-        grantTypeInfo?.authEndpoint,
-        (value) => {
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            authEndpoint: value,
-          }
-        }
-      )
-
-      const clientID = refWithCallbackOnChange(
-        grantTypeInfo?.clientID,
-        (value) => {
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            clientID: value,
-          }
-        }
-      )
-
-      const scopes = refWithCallbackOnChange(
-        grantTypeInfo?.scopes ? grantTypeInfo.scopes : undefined,
-        (value) => {
-          auth.value.grantTypeInfo = {
-            ...auth.value.grantTypeInfo,
-            scopes: value,
-          }
-        }
-      )
-
-      const runAction = () => {
-        const values: ImplicitOauthFlowParams =
-          replaceTemplateStringsInObjectValues({
-            authEndpoint: authEndpoint.value,
-            clientID: clientID.value,
-            scopes: scopes.value,
-          })
-
-        const unwrappedValues = replaceTemplateStringsInObjectValues(values)
-
-        const parsedArgs = implicit.params.safeParse(unwrappedValues)
-
-        if (!parsedArgs.success) {
-          return E.left("VALIDATION_FAILED" as const)
-        }
-
-        implicit.init(parsedArgs.data)
-
-        return E.right(undefined)
-      }
-
-      const elements = computed(() => {
-        return [
-          {
-            id: "authEndpoint",
-            label: t("authorization.oauth.label_authorization_endpoint"),
-            type: "text" as const,
-            ref: authEndpoint,
-          },
-          {
-            id: "clientId",
-            label: t("authorization.oauth.label_client_id"),
-            type: "text" as const,
-            ref: clientID,
-          },
-          {
-            id: "scopes",
-            label: t("authorization.oauth.label_scopes"),
-            type: "text" as const,
-            ref: scopes,
-          },
-        ]
-      })
-
-      return {
-        runAction,
-        elements,
-      }
-    }),
-  },
-]
-
-type GrantTypes = z.infer<
-  typeof HoppRESTAuthOAuth2
->["grantTypeInfo"]["grantType"]
-
-const grantTypeMap: Record<
-  GrantTypes,
-  (typeof supportedGrantTypes)[number]["id"]
-> = {
-  AUTHORIZATION_CODE: "authCode",
-  CLIENT_CREDENTIALS: "clientCredentials",
-  IMPLICIT: "implicit",
-  PASSWORD: "password",
-} as const
-
-const grantTypeDefaultPayload = {
-  AUTHORIZATION_CODE: getDefaultAuthCodeOauthFlowParams,
-  CLIENT_CREDENTIALS: getDefaultClientCredentialsFlowParams,
-  IMPLICIT: getDefaultImplicitOauthFlowParams,
-  PASSWORD: getDefaultPasswordFlowParams,
-} as const
-
-const selectedGrantTypeID = computed(() => {
-  const currentGrantType = auth.value.grantTypeInfo.grantType
-  return grantTypeMap[currentGrantType]
-})
-
-const selectedGrantType = computed(() => {
-  return supportedGrantTypes.find(
-    (grantType) => grantType.id === selectedGrantTypeID.value
-  )
-})
+const gqlTabsService = useService(GQLTabService)
+const restTabsService = useService(RESTTabService)
+const persistenceService = useService(PersistenceService)
 
 const setAccessTokenInActiveContext = (
   accessToken?: string,
@@ -1374,69 +692,48 @@ const setAccessTokenInActiveContext = (
     return
   }
 
-  const tabService = props.source === "REST" ? restTabsService : gqlTabsService
-
-  if (
-    tabService.currentActiveTab.value.document.request.auth.authType ===
-      "oauth-2" &&
-    accessToken
-  ) {
-    tabService.currentActiveTab.value.document.request.auth.grantTypeInfo.token =
+  if (props.source === "REST") {
+    const restTab = restTabsService.currentActiveTab.value
+    if (
+      "request" in restTab.document &&
+      restTab.document.request &&
+      restTab.document.request.auth.authType === "oauth-2" &&
       accessToken
-  }
+    ) {
+      restTab.document.request.auth.grantTypeInfo.token = accessToken
+    }
 
-  if (
-    refreshToken &&
-    tabService.currentActiveTab.value.document.request.auth.authType ===
-      "oauth-2"
-  ) {
-    // @ts-expect-error - TODO: narrow the grantType to only supporting refresh tokens
-    tabService.currentActiveTab.value.document.request.auth.grantTypeInfo.refreshToken =
-      refreshToken
-  }
-}
+    if (
+      refreshToken &&
+      "request" in restTab.document &&
+      restTab.document.request &&
+      restTab.document.request.auth.authType === "oauth-2"
+    ) {
+      // @ts-expect-error - TODO: narrow the grantType to only supporting refresh tokens
+      restTab.document.request.auth.grantTypeInfo.refreshToken = refreshToken
+    }
+  } else {
+    const gqlTab = gqlTabsService.currentActiveTab.value
+    if (
+      "request" in gqlTab.document &&
+      gqlTab.document.request &&
+      gqlTab.document.request.auth.authType === "oauth-2" &&
+      accessToken
+    ) {
+      gqlTab.document.request.auth.grantTypeInfo.token = accessToken
+    }
 
-const changeSelectedGrantType = (
-  grantType: (typeof supportedGrantTypes)[number]["id"]
-) => {
-  const keys = Object.keys(grantTypeMap) as GrantTypes[]
-
-  const grantTypeToSet = keys.find((key) => grantTypeMap[key] === grantType)
-
-  if (grantTypeToSet) {
-    auth.value.grantTypeInfo.grantType = grantTypeToSet
-
-    const getDefaultPayload = grantTypeDefaultPayload[grantTypeToSet]
-
-    // set the default payload for the grant type
-    // for eg: if the grant type was auth code, and then the user selected the password grant type,
-    // there wont be a password key in the payload
-    // so we set the default payload, and it'll add all the keys that are missing
-    if (getDefaultPayload) {
-      auth.value.grantTypeInfo = {
-        ...getDefaultPayload(),
-        ...auth.value.grantTypeInfo,
-      }
+    if (
+      refreshToken &&
+      "request" in gqlTab.document &&
+      gqlTab.document.request &&
+      gqlTab.document.request.auth.authType === "oauth-2"
+    ) {
+      // @ts-expect-error - TODO: narrow the grantType to only supporting refresh tokens
+      gqlTab.document.request.auth.grantTypeInfo.refreshToken = refreshToken
     }
   }
 }
-
-const runAction = computed(() => {
-  return selectedGrantType.value?.formElements.value?.runAction
-})
-
-const runTokenRefresh = computed(() => {
-  // the only grant type that supports refresh tokens is the authCode grant type
-  if (selectedGrantType.value?.id === "authCode") {
-    return selectedGrantType.value?.formElements.value?.refreshToken
-  }
-
-  return null
-})
-
-const currentOAuthGrantTypeFormElements = computed(() => {
-  return selectedGrantType.value?.formElements.value?.elements.value
-})
 
 const refreshOauthToken = async () => {
   if (!runTokenRefresh.value) {
@@ -1446,7 +743,7 @@ const refreshOauthToken = async () => {
   const res = await runTokenRefresh.value()
 
   if (E.isLeft(res)) {
-    const errorMessages = {
+    const errorMessages: Record<string, string> = {
       NO_REFRESH_TOKEN_PRESENT: t(
         "authorization.oauth.no_refresh_token_present"
       ),
@@ -1507,323 +804,4 @@ const generateOAuthToken = async () => {
     return
   }
 }
-
-const getPlaceholderForField = (fieldId: string): string => {
-  const placeholders: Record<string, string> = {
-    authEndpoint: "https://example.com/oauth2/authorize",
-    tokenEndpoint: "https://example.com/oauth2/token",
-    clientId: "your_client_id_here",
-    clientSecret: "your_client_secret_here",
-    scopes: "read write",
-    username: "your_username",
-    password: "your_password",
-  }
-  return placeholders[fieldId] || t("authorization.oauth.enter_value")
-}
-
-const grantTypeTippyActions = ref<HTMLElement | null>(null)
-const pkceTippyActions = ref<HTMLElement | null>(null)
-const authTippyActions = ref<HTMLElement | null>(null)
-const clientAuthenticationTippyActions = ref<HTMLElement | null>(null)
-
-// Advanced Configuration state
-const isAdvancedConfigExpanded = ref(false)
-
-const toggleAdvancedConfig = () => {
-  isAdvancedConfigExpanded.value = !isAdvancedConfigExpanded.value
-}
-
-// Advanced Configuration: Auth Request Parameters
-type OAuth2AdvancedParam = {
-  id: number
-  key: string
-  value: string
-  active: boolean
-  sendIn?: "headers" | "url" | "body"
-}
-
-let paramsIdCounter = 1000
-
-// Initialize working auth request params
-const workingAuthRequestParams = ref<OAuth2AdvancedParam[]>([
-  { id: paramsIdCounter++, key: "", value: "", active: true },
-])
-
-// Watch for changes in working auth request params
-watch(
-  workingAuthRequestParams,
-  (newParams: OAuth2AdvancedParam[]) => {
-    // Auto-add empty row when the last row is filled
-    if (newParams.length > 0 && newParams[newParams.length - 1].key !== "") {
-      workingAuthRequestParams.value.push({
-        id: paramsIdCounter++,
-        key: "",
-        value: "",
-        active: true,
-      })
-    }
-
-    // Update auth.value.grantTypeInfo with non-empty params
-    const nonEmptyParams = newParams.filter(
-      (p: OAuth2AdvancedParam) => p.key !== "" || p.value !== ""
-    )
-
-    if ("authRequestParams" in auth.value.grantTypeInfo) {
-      auth.value.grantTypeInfo.authRequestParams = nonEmptyParams.map(
-        (param) => ({
-          id: param.id,
-          key: param.key,
-          value: param.value,
-          active: param.active,
-        })
-      )
-    }
-  },
-  { deep: true }
-)
-
-// Functions for auth request params management
-const addAuthRequestParam = () => {
-  workingAuthRequestParams.value.push({
-    id: paramsIdCounter++,
-    key: "",
-    value: "",
-    active: true,
-  })
-}
-
-const updateAuthRequestParam = (
-  index: number,
-  payload: OAuth2AdvancedParam
-) => {
-  workingAuthRequestParams.value[index] = payload
-}
-
-const deleteAuthRequestParam = (index: number) => {
-  // Only delete if it's not the last empty row, or if there are multiple rows
-  if (workingAuthRequestParams.value.length > 1) {
-    workingAuthRequestParams.value.splice(index, 1)
-  }
-}
-
-// Token Request Parameters
-interface OAuth2TokenParam {
-  id: number
-  key: string
-  value: string
-  sendIn: "headers" | "body" | "url"
-  active: boolean
-}
-
-// Initialize working token request params
-const workingTokenRequestParams = ref<OAuth2TokenParam[]>([
-  {
-    id: paramsIdCounter++,
-    key: "",
-    value: "",
-    sendIn: "body",
-    active: true,
-  },
-])
-
-// Watch for changes in working token request params
-watch(
-  workingTokenRequestParams,
-  (newParams: OAuth2TokenParam[]) => {
-    // Auto-add empty row when the last row is filled
-    if (newParams.length > 0 && newParams[newParams.length - 1].key !== "") {
-      workingTokenRequestParams.value.push({
-        id: paramsIdCounter++,
-        key: "",
-        value: "",
-        sendIn: "body",
-        active: true,
-      })
-    }
-
-    // Update auth.value.grantTypeInfo with non-empty params
-    const nonEmptyParams = newParams.filter(
-      (p: OAuth2TokenParam) => p.key !== "" || p.value !== ""
-    )
-
-    if ("tokenRequestParams" in auth.value.grantTypeInfo) {
-      auth.value.grantTypeInfo.tokenRequestParams = nonEmptyParams.map(
-        (param) => ({
-          id: param.id,
-          key: param.key,
-          value: param.value,
-          sendIn: param.sendIn,
-          active: param.active,
-        })
-      )
-    }
-  },
-  { deep: true }
-)
-
-// Functions for token request params management
-const addTokenRequestParam = () => {
-  workingTokenRequestParams.value.push({
-    id: paramsIdCounter++,
-    key: "",
-    value: "",
-    sendIn: "body",
-    active: true,
-  })
-}
-
-const updateTokenRequestParam = (index: number, payload: OAuth2TokenParam) => {
-  workingTokenRequestParams.value[index] = payload
-}
-
-const deleteTokenRequestParam = (index: number) => {
-  // Only delete if it's not the last empty row, or if there are multiple rows
-  if (workingTokenRequestParams.value.length > 1) {
-    workingTokenRequestParams.value.splice(index, 1)
-  }
-}
-
-// Refresh Request Parameters
-interface OAuth2RefreshParam {
-  id: number
-  key: string
-  value: string
-  sendIn: "headers" | "body" | "url"
-  active: boolean
-}
-
-// Initialize working refresh request params
-const workingRefreshRequestParams = ref<OAuth2RefreshParam[]>([
-  {
-    id: paramsIdCounter++,
-    key: "",
-    value: "",
-    sendIn: "body",
-    active: true,
-  },
-])
-
-// Watch for changes in working refresh request params
-watch(
-  workingRefreshRequestParams,
-  (newParams: OAuth2RefreshParam[]) => {
-    // Auto-add empty row when the last row is filled
-    if (newParams.length > 0 && newParams[newParams.length - 1].key !== "") {
-      workingRefreshRequestParams.value.push({
-        id: paramsIdCounter++,
-        key: "",
-        value: "",
-        sendIn: "body",
-        active: true,
-      })
-    }
-
-    // Update auth.value.grantTypeInfo with non-empty params
-    const nonEmptyParams = newParams.filter(
-      (p: OAuth2RefreshParam) => p.key !== "" || p.value !== ""
-    )
-
-    if ("refreshRequestParams" in auth.value.grantTypeInfo) {
-      auth.value.grantTypeInfo.refreshRequestParams = nonEmptyParams.map(
-        (param) => ({
-          id: param.id,
-          key: param.key,
-          value: param.value,
-          sendIn: param.sendIn,
-          active: param.active,
-        })
-      )
-    }
-  },
-  { deep: true }
-)
-
-// Functions for refresh request params management
-const addRefreshRequestParam = () => {
-  workingRefreshRequestParams.value.push({
-    id: paramsIdCounter++,
-    key: "",
-    value: "",
-    sendIn: "body",
-    active: true,
-  })
-}
-
-const updateRefreshRequestParam = (
-  index: number,
-  payload: OAuth2RefreshParam
-) => {
-  workingRefreshRequestParams.value[index] = payload
-}
-
-const deleteRefreshRequestParam = (index: number) => {
-  // Only delete if it's not the last empty row, or if there are multiple rows
-  if (workingRefreshRequestParams.value.length > 1) {
-    workingRefreshRequestParams.value.splice(index, 1)
-  }
-}
-
-// Initialize advanced parameters from the auth object when component mounts
-onMounted(() => {
-  if (
-    "authRequestParams" in auth.value.grantTypeInfo &&
-    auth.value.grantTypeInfo.authRequestParams &&
-    auth.value.grantTypeInfo.authRequestParams.length > 0
-  ) {
-    workingAuthRequestParams.value =
-      auth.value.grantTypeInfo.authRequestParams.map((param) => ({
-        id: param.id || paramsIdCounter++,
-        key: param.key,
-        value: param.value,
-        active: param.active,
-      }))
-  }
-
-  if (
-    "tokenRequestParams" in auth.value.grantTypeInfo &&
-    auth.value.grantTypeInfo.tokenRequestParams &&
-    auth.value.grantTypeInfo.tokenRequestParams.length > 0
-  ) {
-    workingTokenRequestParams.value = [
-      ...auth.value.grantTypeInfo.tokenRequestParams.map((param) => ({
-        id: param.id || paramsIdCounter++,
-        key: param.key,
-        value: param.value,
-        sendIn: param.sendIn || "body",
-        active: param.active,
-      })),
-
-      {
-        id: paramsIdCounter++,
-        key: "",
-        value: "",
-        sendIn: "body",
-        active: true,
-      },
-    ]
-  }
-
-  if (
-    "refreshRequestParams" in auth.value.grantTypeInfo &&
-    auth.value.grantTypeInfo.refreshRequestParams &&
-    auth.value.grantTypeInfo.refreshRequestParams.length > 0
-  ) {
-    workingRefreshRequestParams.value = [
-      ...auth.value.grantTypeInfo.refreshRequestParams.map((param) => ({
-        id: param.id || paramsIdCounter++,
-        key: param.key,
-        value: param.value,
-        sendIn: param.sendIn || "body",
-        active: param.active,
-      })),
-      {
-        id: paramsIdCounter++,
-        key: "",
-        value: "",
-        sendIn: "body",
-        active: true,
-      },
-    ]
-  }
-})
 </script>
